Add methods from Rect
